// src/types.d.ts
export interface Staff {
    name: string, 
    lastName: string, 
    age: number, 
    job: string, 
    active: boolean
  }


  export interface ResultMovies {
    "adult": boolean,
    "backdrop_path": string,
    "genre_ids": Array<number>,
    "id": number,
    "original_language": string,
    "original_title": string,
    "overview": string,
    "popularity": number,
    "poster_path": string,
    "release_date": string,
    "title": string,
    "video": boolean,
    "vote_average": number,
    "vote_count": number
}

export interface Movies {
    "page": number,
    "results": Array<ResultMovies>}